Move all target-specific code to harness.rs
