+++ fuzzers/binary_only/qemu_launcher/README.md
@@ -1,9 +1,13 @@
 # qemu_launcher_
 
-This folder contains an example fuzzer for libpng, using LLMP for fast multi-process fuzzing and crash detection. It has been tested on Linux.
-This automatically spawns n child processes, and binds them to a free core.
+This folder contains an example fuzzer that will fuzz binary-only targets, cross-architecture, on Linux.
+It's using LLMP for fast multi-process fuzzing and crash detection.
+This automatically spawns `n` child processes, and binds them to a free core.
+
+To adapt the fuzzer to your custom target, change [`harness.rs`](./src/harness.rs).
 
 The following architectures are supported:
+
 * arm
 * aarch64
 * i386
@@ -11,10 +15,10 @@ The following architectures are supported:
 * mips
 * ppc
 
-Note that the injection feature `-j` is currently only supported on x86_64
-and aarch64.
+For usermode, this fuzzer supports injection fuzzing with `-j`.
 
 ## Prerequisites
+
 ```bash
 sudo apt install \
     gcc-arm-linux-gnueabi \
@@ -32,7 +36,8 @@ sudo apt install \
 
 ## Run
 
-Defaults to `x86_64` architecture
+Defaults to `x86_64` architecture. Change the architecture by 
+
 ```bash
 cargo make run
 ```
