+++ fuzzers/binary_only/qemu_launcher/src/client.rs
@@ -11,16 +11,16 @@ use libafl_bolts::{core_affinity::CoreId, rands::StdRand, tuples::tuple_list};
 #[cfg(feature = "injections")]
 use libafl_qemu::modules::injections::InjectionModule;
 use libafl_qemu::{
-    elf::EasyElf,
     modules::{
         asan::{init_qemu_with_asan, AsanModule},
         asan_guest::{init_qemu_with_asan_guest, AsanGuestModule},
         cmplog::CmpLogModule,
     },
-    ArchExtras, GuestAddr, Qemu,
+    Qemu,
 };
 
 use crate::{
+    harness::Harness,
     instance::{ClientMgr, Instance},
     options::FuzzerOptions,
 };
@@ -55,16 +55,7 @@ impl<'a> Client<'a> {
             .collect::<Vec<(String, String)>>()
     }
 
-    fn start_pc(qemu: Qemu) -> Result<GuestAddr, Error> {
-        let mut elf_buffer = Vec::new();
-        let elf = EasyElf::from_file(qemu.binary_path(), &mut elf_buffer)?;
-
-        let start_pc = elf
-            .resolve_symbol("LLVMFuzzerTestOneInput", qemu.load_addr())
-            .ok_or_else(|| Error::empty_optional("Symbol LLVMFuzzerTestOneInput not found"))?;
-        Ok(start_pc)
-    }
-
+    #[allow(clippy::too_many_lines)]
     pub fn run<M: Monitor>(
         &self,
         state: Option<ClientState>,
@@ -72,9 +63,11 @@ impl<'a> Client<'a> {
         core_id: CoreId,
     ) -> Result<(), Error> {
         let mut args = self.args()?;
+        Harness::edit_args(&mut args);
         log::debug!("ARGS: {:#?}", args);
 
         let mut env = self.env();
+        Harness::edit_env(&mut env);
         log::debug!("ENV: {:#?}", env);
 
         let is_asan = self.options.is_asan_core(core_id);
@@ -96,9 +89,6 @@ impl<'a> Client<'a> {
             }
         };
 
-        let start_pc = Self::start_pc(qemu)?;
-        log::debug!("start_pc @ {start_pc:#x}");
-
         #[cfg(not(feature = "injections"))]
         let injection_module = None;
 
@@ -118,13 +108,7 @@ impl<'a> Client<'a> {
                 }
             });
 
-        qemu.entry_break(start_pc);
-
-        let ret_addr: GuestAddr = qemu
-            .read_return_address()
-            .map_err(|e| Error::unknown(format!("Failed to read return address: {e:?}")))?;
-        log::debug!("ret_addr = {ret_addr:#x}");
-        qemu.set_breakpoint(ret_addr);
+        let harness = Harness::init(qemu).expect("Error setting up harness.");
 
         let is_cmplog = self.options.is_cmplog_core(core_id);
 
@@ -136,6 +120,7 @@ impl<'a> Client<'a> {
         let instance_builder = Instance::builder()
             .options(self.options)
             .qemu(qemu)
+            .harness(harness)
             .mgr(mgr)
             .core_id(core_id)
             .extra_tokens(extra_tokens);
